Added ReporterFileVisitor outside of test files
- Cleaned up tests a bit.
- Started working on transformer which ensures reporter import.
- Added a test case class which sets up temp directory as required for
tests which modify code.
